added ViewBinding, NavigationGraph. The start screen has been moved to a fragment.
