+++ app/build.gradle.kts
@@ -49,9 +49,9 @@ android {
         jvmTarget = "17"
     }
 
-//    buildFeatures {
-//        viewBinding = true
-//    }
+    buildFeatures {
+        viewBinding = true
+    }
 }
 
 dependencies {
@@ -75,6 +75,10 @@ dependencies {
 
     // Room
     implementation(libs.room.runtime)
+    implementation(libs.androidx.legacy.support.v4)
+    implementation(libs.androidx.lifecycle.livedata.ktx)
+    implementation(libs.androidx.lifecycle.viewmodel.ktx)
+    implementation(libs.androidx.fragment.ktx)
     ksp(libs.room.compiler)
     implementation(libs.room.ktx)
