+++ app/src/main/java/com/tinkoff/android_homework/presentation/detail/DetailFragment.kt
@@ -2,21 +2,38 @@ package com.tinkoff.android_homework.presentation.detail
 
 import android.os.Bundle
 import android.util.Log
+import android.view.LayoutInflater
 import android.view.View
+import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
+import com.tinkoff.android_homework.databinding.FragmentDetailBinding
+import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 
 /**
  * @author d.shtaynmets
  */
-class DetailFragment : Fragment() {
 
+@AndroidEntryPoint
+class DetailFragment : Fragment() {
     private val viewModel by viewModels<DetailViewModel>()
+
+    private var binding: FragmentDetailBinding? = null
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = FragmentDetailBinding.inflate(inflater, container, false)
+
+        return binding!!.root
+    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
         lifecycleScope.launch {
             Log.e("TAGRTRT", "viewModel.details.value :${viewModel.details.value}")
         }
