+++ gradle/libs.versions.toml
@@ -19,6 +19,10 @@ ksp = "2.0.0-1.0.22"
 nav_graph = "2.8.6"
 hilt-version = "2.51.1"
 view-binding-version = "8.8.0"
+legacy-support-v4 = "1.0.0"
+lifecycle-livedata-ktx = "2.8.7"
+lifecycle-viewmodel-ktx = "2.8.7"
+fragment-ktx = "1.8.5"
 
 [libraries]
 core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "core-ktx" }
@@ -43,6 +47,10 @@ navigation-graph-ui = { group = "androidx.navigation", name = "navigation-ui-ktx
 hilt-android = { group = "com.google.dagger", name = "hilt-android", version.ref = "hilt-version" }
 hilt-compiler = { group = "com.google.dagger", name = "hilt-android-compiler", version.ref = "hilt-version" }
 view-binding = { group = "com.android.databinding", name = "viewbinding", version.ref = "view-binding-version" }
+androidx-legacy-support-v4 = { group = "androidx.legacy", name = "legacy-support-v4", version.ref = "legacy-support-v4" }
+androidx-lifecycle-livedata-ktx = { group = "androidx.lifecycle", name = "lifecycle-livedata-ktx", version.ref = "lifecycle-livedata-ktx" }
+androidx-lifecycle-viewmodel-ktx = { group = "androidx.lifecycle", name = "lifecycle-viewmodel-ktx", version.ref = "lifecycle-viewmodel-ktx" }
+androidx-fragment-ktx = { group = "androidx.fragment", name = "fragment-ktx", version.ref = "fragment-ktx" }
 
 #hilt-compiler = { group = "com.google.dagger", name = "hilt-compiler", version.ref = "hilt" }
 #hilt-android = { group = "com.google.dagger", name = "hilt-android", version.ref = "hilt" }
